+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - Add style to route if on route
 
 ## 0.2.1
-- Fix order of foudn herbs
+- Fix order of found herbs
 
 ## 0.2.0
 - Herbalism Generator added